:file_folder: Move routes to app/routes
