+++ app/routes/index.js
@@ -4,24 +4,24 @@ var fs = require('fs');
 var dbName = "Jakhu";
 var port = "27017";
 var host = "localhost";
-var kernal = require('../app/boot/boot.js');
+var kernal = require('../boot/boot.js');
 var exec = require('child_process').exec;
 var xml2js = require('xml2js');
-var oobe = require('../libs/setup/setup.js');
+var oobe = require('../../libs/setup/setup.js');
 var os = require('os');
 var clicolour = require('cli-color');
-var db = require('../libs/database');
+var db = require('../../libs/database');
 var app = express();
 var expresssession = require('express-session');
 var passport = require('passport');
 var LocalStrategy = require('passport-local').Strategy;
 var mongoose = require('mongoose');
 var LocalStrategy = require('passport-local').Strategy;
-var User = require('../libs/database').user;
+var User = require('../../libs/database').user;
 var bcrypt = require('bcryptjs');
-var plocal = require('../libs/passport/local.js');
-var pgit = require('../libs/passport/github.js');
-var prem = require('../libs/passport/rember.js');
+var plocal = require('../../libs/passport/local.js');
+var pgit = require('../../libs/passport/github.js');
+var prem = require('../../libs/passport/rember.js');
 var utils = require('util');
 
 // Passport session setup.
