+++ config/config.xml
@@ -1,1 +1,5 @@
-ECHO is on.
+<config>
+    <language>english</language>
+    <region>uk</region>
+    <allowSend>true</allowSend>
+</config>
